Oops.  Goof in my last commit, which I apparently didn't test well enough. Caught by tedkaz@optonline.net (Ted Kaczmarek) and noted in Bug #961:
	http://bugzilla.opennms.org/cgi-bin/bugzilla/show_bug.cgi?id=961

STATUS_WAIT was supposed to be set to 5, but it was set to 0.
The startup script would check 10 times (START_TIMEOUT) to see if all
OpenNMS services had started, but it would do them all in a very short
period of time, almost always before all of the OpenNMS services could
start up.  It's set back to 5 now, so the script will wait up to 50
seconds for OpenNMS to start.