Make the jsp.compile target work with the new jasper libraries that were updated a day or two ago.
The new libraries set the package name in the .java file they create to be
the path under the uriroot passed to the jasper2 task, excluding the name
ofthe .jsp file.  The old libraries always used org.apache.jsp.  We used
to compile each .java file individually into a temporary directory and then
move the resulting class into a directory structure like that of the source
JSP files.  The new jasper library effectively does this for us by setting
the package name.  This is sweet, because it lets us get rid of the for
loop from the jsp.compile.javac target and compile everything in a single
call to javac.  Much faster!  w00t!