Use a small Java app to get status information, so we don't have to depend on parse-status.pl and curl/wget/etc.  It's also faster than calling the real manager app (which opennnms.sh would fall back to).
It's not as fast as I like ("opennms status" takes 1.5 seconds on my
Mac OS X box), but it isn't much slower than the previous solution
and it should make opennms.sh functions more reliable.

This also opens up the door to requiring authentication to the JMX
server in OpenNMS, which would be good thing, to say the least. ;)