First pass of two or three thru the service monitors
In this pass I removed all the deprecated checkStatus methods and replaced them
with the officially sanctioned poll method.  Made each return a PollStatus object
and made an attempt at added the appropriate Reason codes for failure cases.

Also attempted to ensure the responseTime in included in all Available status