Boog 1644: Fixed testChangeIpAddr.
The problem here was that we were running multiple importers in the same JVM on the same thread
and some of the caching code in the importer was using a static ThreadLocal so it though all
the categories were already there.  Changed these to a new ThreadLocal per importer.

Also turned down the logging