a little refactoring of the jettyserver
I was hoping to make it so you could define a "default" context (ie, "opennms") and have it listen on both /context and / (so the opennms webapp would also be available as the "ROOT").

But no go.

Maybe we can make a small redirector context instead and have it forward all requests to 'defaultcontext's url.