Approved by tarus/jeffg/david(?)@
a) Add a "reloadCheckInterval" property that specifies the minimum amount of
time between checking to see if the underlying file resource (if any) has been
updated.  This is to keep us from stat()ing left and right when a lot of calls
are made to getObject().  The default is 1000 millieconds, 0 forces the check
to always happen when getObject() is called, and a negative number disables
auto-refresh.  The last option was added to support DefaultEventConfDao, which
doesn't auto-refresh.
b) Add a reload() method so we can be explicitly reloaded.  Added in support of
DefaultEventConfDao's reload needs.
c) Add a description to AbstractCastorConfigDao to enhance log and exception
messages.
d) Add reload check interval support to AbstractCastorConfigDao, delegating to
FileReloadContainer.
e) Log the amount of time it took to load a configuration in
AbstractCastorConfigDao, allowing the log message to be overridden in
subclasses that wish to log more detail (this is to support a feature that came
from DefaultEventConfDao that I didn't want to lose).
f) Change all of the subclasses of AbstractCastorConfigDao to pass the
newly-required description argument to their super constructor.