Fixed bugs 3446 and 3488.
For bug 3446, The JdbcIfLabel dependency was removed from
CollectionResourceWrapper, only for SNMP interface resources (i.e.
resources comming from collectd). For latency resources (i.e. comming
from pollerd), the code will use JdbcIfLabel to obtain interface label
and ifIndex.

The main problem reported for bug 3446 was for thresholds on collectd,
so the solution will fix that.

The solution was, emulate the content from snmpinterface table using
OnmsSnmpInterface on SnmpIfData and then use it on IfInfo.

For bug 3488, the solution was more simple, just validate that the
interface label was successfuly retrieved on CollectionResourceWrapper
before trying to read ifInfo.