Fixed bug #3619: Refactored SnmpInterfacePoller
The object Model was revisited. Now The annotation Listener is used.
Also the Poller uses the snmpinterface table data for initialization.
The events are sent only when a state change occurs.
Only one event is sent for state change. The alarm definition is changed
according with this rule using the same reduction key for Oper and Admin related events.