Fix for bugs 3427 and 3428
Bug 3427:

The class AbstractThresholdEvaluatorState will format the threshold
value using java.text.DecimalFormat, based on the system property:

org.opennms.threshd.value.decimalformat.

The default value for this property is '###.##', which means that the
threshold value will show only 2 decimals. If you want to override this,
add the property to the opennms.properties file.

Bug 3428:

The file thresholding-config.xsd was modified to include a new
attribute on basethresholddef called "filterOperator". This attribute
has two possible values: "or", "and". The attribute is optional, and if
it's not  supplied, the default value will be "or" (because this is the
current default behavior for resource-filter evaluation).

If you specify filterOperator with the word "and" as a value, then ALL
resource filters should match in order to trigger the threshold. For
example:

<threshold type="high" ds-name="ifInOctets"  ds-type="if" value="90"
    rearm="50" trigger="1" filterOperator="and">
  <resource-filter field='snmpifdescr'>wlan0</resource-filter>
  <resource-filter field='snmpifspeed'>100000000</resource-filter>
</threshold>