Fixed bug 3711: use browser windows size to size the map.
The maps now use the browser windows size to set the window dimension of the map.
A minimum map size of 1272x600 is set also if the browser windows size is less.
Added a window.resize function to capture resize events and resetting the map
size according with the broser size.