fix bug #3989, incorrect bounds-checking on insert
When inserting initial vlan data, if certain conditions aren't met, it would not
append all columns to the insert, which would result in a null vlan name.

This bounds-checking is only appropriate on update, when it detects what has been
modified since the last commit to the vlan table.

Also, update some of the toString() methods on the vlan entry objects to be easier
to find when logging.