SPC-120 - send events when POSTing to ReST
When the NCS ReST API makes any modifications to the components
in the database, it will send events about which components
were updated.  In addition, this fixes the schema so a given
type/foreignSource/foreignId is unique, and cleans up any existing
data so only one of each exists (keeps the newest, reparents
anything that refers to old components).