NMS-5514 Reverse DNS Adapter
this adapter update reverse DNS zone with the data holded in opennms
The reverse zone is updated using a dns PTR record with value {interfaceName}-{FQHN}.
This is really useful in traceroute so that you are able to get the name of the node and the interface too.

This adapter will work in the same way as the DNS Provisiong Adapter.
This adapter support only ipv4 third level Reverse DNS Zone.