Squash merge of name-cutter branch
A new project called name-cutter has been added by extracting the
required classes from the opennms-jmx-tools.

These classes provide the ability to shrink names in order to avoid
problems with the 19 characters restriction for the DS names on JRB/RRD
files.

Now, the opennms-jmx-tools and vaadin-snmp-events-and-metrics are using
this library to generate the configuration files.

Also, the datacollection configuration UI has been changed to display
basic validations. These validations are just for visual purposes
because in order to propagate the validations to avoid save the changes
when there are errors, the code must be refactored.