New Feature: Current Problems Box (BD-2)
This is an optional box that can be displayed on the frontpage that
contains information about the current nodes with problems based on the
outstanding alarms (i.e. not acknowledged) that has a severity greater
than 3, classified by node, and ordered by the oldest alarm, on a
similar way to the outage box. Also each row will have the color of the
greatest severity of the alarms processed on each case.

This box is more generic and has more information than the outage box,
as outages are also alarms.

This feature is disabled by default but can be enabled in
opennms.properties:

This value allows you to set the number of nodes with outages to
display on the front page in the OpenNMS web UI. Default 12

opennms.nodesWithOutages.count=12

This value allows you to enable/disable the nodes with outages box on
the front page in the OpenNMS web UI. Default: true

opennms.nodesWithOutages.show=true

This value allows you to set the number of nodes with problems to
display on the front page in the OpenNMS web UI. Default: 12

opennms.nodesWithProblems.count=12

This value allows you to enable/disable the nodes with problems box on
the front page in the OpenNMS web UI. Default: false

opennms.nodesWithProblems.show=false

This value allows you to enable/disable the status bar resume at the
top of the node page in the OpenNMS web UI. Default: false

opennms.nodeStatusBar.show=false

The last option is related with other new optional feature, the node
status bar that can be displayed at the top of the node's page
containing the greatest severity of the alarm the node has, with other
useful information.