Clean-up MIB Tree Container and Removing Filename Restriction.
There is no need to define and keep a reference to a
HierarchicalContainer. The Tree object should handle
everything.

Also the comment at the top that says:
"The file name requires to be the same as the MIB to
be processed."
Is not necessary anymore, because the MIB Parser will
rename the file to be the name of the MIB automatically.