Fix for Bug NMS-6072 - Part II
The initial solution was to check if the response is a redirect. If so,
perform a head request to verify the location header. The problem is
that the HEAD operation will be treated as a GET operation, which is
going to force the synchronous operation of creating the file on the
pending directory which is not correct.

Now, if the response is a redirect provision.pl will consider it correct
if there is a location header and that's it.