Applying the fix for NMS-4902 on PhysInterfaceTableTracker.
The way to process the ifPhysAddress was changed in IfTableEntry because
of NMS-4902 (added the MAC address handling from IpNetToMediaTableEntry
to IfTableEntry). The fix for NMS-5418 didn't include the same
post-processing of the MAC Address which is not correct.

Now both clases are going to provide the same functionality regarding
the MAC address.