Fix for NMS-6140 - Rule for selecting the primary interface on ESX
This fix contains the hability to resolve the ESX Name as FQDN, and use
it as the primary interface if that IP is listed on the IPs associated
with the HostSystem, otherwise use the old rule (use the first IP).

Pending:

The code still need to apply the correct behavior when the FQDN solution
doesn't work which is use the IP with the port property called
"Management Traffic" enabled on the "Management Network".