Fix for NMS-6359 - Issues with store-by-FS and generic resources only.
The lazy loader used in NodeResourceType was assuming the old convention
when searching for child resources. This loader have been refactored in
order to use the proper methods from the OnmsResourceType
implementations to obtain the list of children.

Added a method called ResourceTypeUtils.isStoreByForeignSource to avoid
checking for org.opennms.rrd.storeByForeignSource from several places.