Fix for NMS-6409 - Upgrade 1.12.4 breaks StoreByForeignSource
There was a typo error on ResourceTypeUtils. This class was updated in
order to have a centralized way to ask if storeByForeignSource is
enabled (like storeByGroup).

DefaultCollectionAgentService was updated to use ResourceTypeUtils for
this purpose.

Some code was removed from CollectorThresholdingSet related to this
feature because it is not required.