Partial fix for NMS-6386
The ILR reports wrong data when there are several packages with
different collection rates on collectd-configuration.xml for the same
service.

The partial solution is: move the EJN package from the default
collectd-configuration.xml to the one on the examples directory, because
that package is confusing ILR and it was created only for a very
particular kind of devices.