Fix for NMS-6403 - Security Information disclosed in Service detail screen
Now, for custom monitors like PSM or MTM, if there is an attribute
called password or user-info, its content will be masked with "XXXX".
If there is a parameter where the key attribute contains password or
passwd, the value will be masked with "XXXX".

For standard parameters, if the parameter's key contains the word
password, it won't be displayed.

The solution was verified on a testing environment.