Fix for Bug NMS-6489
Upgrade Tools issue: the snmpStorageFlag from datacollection-config.xml
is not being considered while running SnmpInterfaceRrdMigratorOnline.

The filter for selecting the SNMP interfaces where snmpCollect is 'C' or
'UC' was removed in order to consider all kind of scenarios for data
collection, specially when snmpCollectFlag="all".

The solution was tested on a lab using a real DB/RRD from a customer.