Show the delivery options and report parameters for scheduled report jobs.
On Reports -> Database Reports -> Manage the batch report schedule, a
user was able to see the trigger name, the next fire time, and a check
box to unschedule the job.

Now, the user will be able to see the internal report ID, the delivery
options, and the report parameters (if exist) to have a better idea
about the report jobs that have been scheduled and how they have been
configured.