Enhancements for the notifications auto_acknowledge
Notice: the following features are optional, they are going to be used
only if they are configured.

1) Matching event parameters

The original implementaion of auto-acknowledge provides a way to match
only the nodeId, the ipAddress and the serviceName. Now, it is possible
to match any event parameter, for example:

<auto-acknowledge resolution-prefix="RESOLVED: "
    uei="uei.opennms.org/nodes/snmp/interfaceOperUp"
    acknowledge="uei.opennms.org/nodes/snmp/interfaceOperDown">
    <match>nodeid</match>
    <match>interfaceid</match>
    <match>serviceid</match>
    <match>parm[snmpifname]</match>
</auto-acknowledge>

It is possible to provide either the parameter's name (like snmpifname)
or the parameter's number (like parm[#2])

2) Matching alarm's clear-key/reduction-key pairs:

2.1) Global configuration:

<auto-acknowledge-alarm resolution-prefix="RESOLVED: "/>

The above means, all the notifications related with events that have
alarm-data on their configuration are going to be auto-acknowledge.

If this setting is used, all the default entries for auto-acknowledge
can be removed as they are redundant.

2.2) Specific configuration:

<auto-acknowledge-alarm resolution-prefix="RESOLVED: ">
  <uei>uei.opennms.org/nodes/nodeUp</uei>
  <uei>uei.opennms.org/nodes/interfaceUp</uei>
  <uei>uei.opennms.org/nodes/nodeRegainedService</uei>
</auto-acknowledge-alarm>

The above means, apply auto-acknowledge for the provided resolving
events. In this case, if those events comes in, and the originator
events (matching the reduction-key/clear-key pairs) have notifications
configured, those are going to be auto-acknowledged and the resolved
notifications will be sent.