Be more verbose on Statsd's logging
Now you can see the data that potentially is going to be persisted, and
the data that it is actually going to be persisted.

Also, it is going to show a warning when there is no data. If all the
data to be considered for the report is NaN, the report is going to be
empty (and should not be persisted).